home,about,activitiesのレスポンシブ完了とcomponentからpagesにimportする形に変更
